Configure authentication
- Add routes for LDAP authentication and JWT refresh/invalidate
- Configure TTL for JWT and Refresh tokens
- Configure LDAP user provider
- Configure Symfony firewall to expect JWT or LDAP auth

The `LexikJWTAuthentication` bundle handles extracting the JWT token from
the `Authentication: Bearer <token>` header. If present and valid, the
user is authenticated.

The `JWTRefreshToken` bundle add a refresh token and provides the logic
to get a new JWT token when presenting a RefreshToken, and to invalidate
a given RefreshToken.

The Symfony LDAP bundle provides the means to use LDAP authentication in
order to get a new JWT+Refresh token.